Implementation of Mobiletech recurrent payment gateway
- Breaking change into `RecurrentPaymentInterface` was introduced.
Method `charge()` is now required to return state of the charge attempt
to indicate whether it was completed, or just requested and it
remains pending. All pending charges are expected to be confirmed
asynchronously later.
- Added Mobiletech recurrent gateway to charge users via sms.
Payments are confirmed asynchronously once the delivery is confirmed.
If the delivery is not confirmed within 15 minutes, payment is timeouted
and fails via hermes handler.
- All Mobiletech messages now have expiration time set to 100 seconds
(online payments and info messages) and 600 seconds (recurring charges)
so the delivery can be expired on the operator level as well.
- Added `OperatorTypeResolver` capable of determining operator type
based on the phone number prefix or previously received/sent messages
for given phone number.

remp/crm#1197