Fix Stripe 3DS flow in case 3D secure is involved
The original implementation was trigerring 3DS verification
before the window was even submitted - stripe.js trigerred that
automatically when creating setup intent.

This caused that some banks displayed zero amount at the
verification page which is not very OK. At all to be honest.

Setup intent was not actually required to exist and it was
only used to get paymentMethod ID. Creating just the PaymentMethod
with the offline use support is enough.

remp/crm#1893